+++ tests/test_06_math/test_649_nurbs_python_interface.py
@@ -7,7 +7,7 @@ from ezdxf.math import BSpline, global_bspline_interpolation
 
 def test_from_geomdl_curve_to_ezdxf_bspline():
     curve = interpolate_curve([(0, 0), (0, 10), (10, 10), (10, 0)], degree=3)
-    bspline = BSpline.from_geomdl_curve(curve)
+    bspline = BSpline.from_nurbs_python_curve(curve)
     assert bspline.degree == 3
     assert len(bspline.control_points) == 4
     assert len(bspline.knots()) == 8  # count + order
@@ -15,7 +15,7 @@ def test_from_geomdl_curve_to_ezdxf_bspline():
 
 def test_from_ezdxf_bspline_to_geomdl_curve():
     bspline = global_bspline_interpolation([(0, 0), (0, 10), (10, 10), (10, 0)], degree=3)
-    curve = bspline.geomdl_curve()
+    curve = bspline.to_nurbs_python_curve()
     assert curve.degree == 3
     assert len(curve.ctrlpts) == 4
     assert len(curve.knotvector) == 8  # count + order